pyproject.toml
[project]
name = "pln"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "croniter>=3.0.3",
    "llama-index-embeddings-huggingface>=0.3.1",
    "llama-index-readers-web>=0.2.3",
    "llama-index>=0.11.14",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "torch>=2.4.1",
]
